+++ src/main/Waypoint.java
@@ -6,6 +6,7 @@ public abstract class Waypoint {
 
 	private String identifier;
 	private int exitHdg; // In deg
+	private int altFromExtrema; // In ft
 	private double x; // In nm
 	private double y; // In nm
 	
@@ -13,6 +14,11 @@ public abstract class Waypoint {
 	public Waypoint(String identifier, int exitHdg, double x, double y) {
 		this.identifier = identifier;
 		this.exitHdg = exitHdg;
+		int maxAlt = 4000;
+		int minAlt = 0;
+		int alt = (int) (Math.random() * (maxAlt - minAlt + 1)) + minAlt;
+		// Allow clearance when altitude is <= 2000 ft for landing or == max - alt for exiting
+		this.altFromExtrema = this instanceof Runway ? 2000 : AircraftMath.round(alt, Aircraft.ALT_INTERVAL);
 		this.x = x;
 		this.y = y;
 	}
@@ -45,6 +51,27 @@ public abstract class Waypoint {
 
 	public double getTargetY() {
 		return this.y;
+	}
+
+
+	public boolean inRange(double aircraftX, double aircraftY) {
+		double error = 0.5;
+		return aircraftX >= this.getX() - error && aircraftX <= this.getX() + error &&
+			   aircraftY >= this.getY() - error && aircraftY <= this.getY() + error;
+	}
+
+
+	public boolean inTargetRange(double aircraftX, double aircraftY) {
+		double error = 0.5;
+		return aircraftX >= this.getTargetX() - error && aircraftX <= this.getTargetX() + error &&
+			   aircraftY >= this.getTargetY() - error && aircraftY <= this.getTargetY() + error;
+	}
+
+
+	public boolean atAlt(double aircraftAlt, double aircraftMaxAlt) {
+	    if (this instanceof Runway)
+			return aircraftAlt <= this.altFromExtrema;
+		return (int) aircraftAlt == aircraftMaxAlt - this.altFromExtrema;
 	}