+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 ### Fixed
 ### Changed
 
+## [0.1.3] - 3/19/23
+### Added
+* Aircraft must now meet an altitude requirement to be given clearance
+### Fixed
+* Fixed graphical issues involving sin/cos
+* Landing aircraft constantly update heading to point towards runway once they reach the target point
+### Changed
+
 ## [0.1.2] - 3/18/23
 ### Added
 * Landing aircraft will be guided to the end of the approach path then handed off to tower control